video_rag_pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "video-rag-pipeline"
version = "0.1.0"
description = "A robust video processing pipeline that converts MP4 files into LLM-ready JSONL chunks"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Video RAG Pipeline Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    
    # Transcription
    "faster-whisper>=1.0.0",
    
    # Scene detection
    "scenedetect[opencv]>=0.6",
    
    # Image processing
    "opencv-python-headless>=4.8",
    "pillow>=10.0",
    "numpy>=1.24",
    
    # OCR
    "pytesseract>=0.3.10",
    
    # Captioning (BLIP)
    "transformers>=4.35",
    "torch>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-mock>=3.11",
    "ruff>=0.1",
    "mypy>=1.5",
]

[project.scripts]
vrag = "run_pipeline:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["vrag*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
